+++ tests/Functional/Command/Cache/DoctrineResultCacheWarmerCommandTest.php
@@ -25,11 +25,16 @@ namespace App\Tests\Functional\Command\Cache;
 use App\Command\Cache\DoctrineResultCacheWarmerCommand;
 use App\Entity\User;
 use App\Generator\UserCacheIdGenerator;
+use App\Repository\UserRepository;
 use App\Tests\Traits\DatabaseTestingTrait;
+use App\Tests\Traits\LoggerTestingTrait;
 use Doctrine\Common\Cache\Cache;
+use Doctrine\Common\Cache\VoidCache;
+use Doctrine\ORM\Configuration;
 use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\ORM\Query\ResultSetMapping;
 use InvalidArgumentException;
+use Monolog\Logger;
 use Symfony\Bundle\FrameworkBundle\Console\Application;
 use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
 use Symfony\Component\Console\Tester\CommandTester;
@@ -37,12 +42,16 @@ use Symfony\Component\Console\Tester\CommandTester;
 class DoctrineResultCacheWarmerCommandTest extends KernelTestCase
 {
     use DatabaseTestingTrait;
+    use LoggerTestingTrait;
 
     /** @var CommandTester */
     private $commandTester;
 
+    /** @var Configuration */
+    private $ormConfiguration;
+
     /** @var Cache */
-    private $doctrineResultCache;
+    private $resultCacheImplementation;
 
     /** @var UserCacheIdGenerator */
     private $userCacheIdGenerator;
@@ -58,12 +67,39 @@ class DoctrineResultCacheWarmerCommandTest extends KernelTestCase
 
         /** @var EntityManagerInterface $entityManager */
         $entityManager = self::$container->get(EntityManagerInterface::class);
-        $this->doctrineResultCache = $entityManager->getConfiguration()->getResultCacheImpl();
+        $this->ormConfiguration = $entityManager->getConfiguration();
+        $this->resultCacheImplementation = $this->ormConfiguration->getResultCacheImpl();
 
         $this->userCacheIdGenerator = self::$container->get(UserCacheIdGenerator::class);
 
         $this->setUpDatabase();
+        $this->setUpTestLogHandler('cache_warmup');
         $this->loadFixtureByFilename('100usersWithAssignments.yml');
+    }
+
+    public function testItLogsIfUserCacheWarmupWasUnsuccessful(): void
+    {
+        $this->ormConfiguration->setResultCacheImpl(new VoidCache());
+
+        self::assertEquals(0, $this->commandTester->execute(
+            [
+                '--batch-size' => '1',
+            ],
+            [
+                'capture_stderr_separately' => true,
+            ]
+        ));
+
+        for ($i = 1; $i <= 100; $i++) {
+            $username = sprintf('user_%d', $i);
+            $expectedLogMessage = sprintf(
+                "Unsuccessful cache warmup for user '%s' (cache id: '%s')",
+                $username,
+                $this->userCacheIdGenerator->generate($username)
+            );
+
+            $this->assertHasLogRecordWithMessage($expectedLogMessage, Logger::ERROR);
+        }
     }
 
     public function testItCanWarmResultCacheForAllUsers(): void
@@ -81,7 +117,7 @@ class DoctrineResultCacheWarmerCommandTest extends KernelTestCase
             $username = sprintf('user_%d', $i);
             $userCacheId = $this->userCacheIdGenerator->generate($username);
 
-            $this->assertTrue($this->doctrineResultCache->contains($userCacheId));
+            $this->assertTrue($this->resultCacheImplementation->contains($userCacheId));
         }
 
         $this->assertStringContainsString(
@@ -130,14 +166,14 @@ class DoctrineResultCacheWarmerCommandTest extends KernelTestCase
             $username = sprintf('user_%d', $i);
             $userCacheId = $this->userCacheIdGenerator->generate($username);
 
-            $this->assertTrue($this->doctrineResultCache->contains($userCacheId));
+            $this->assertTrue($this->resultCacheImplementation->contains($userCacheId));
         }
 
         for ($i = 11; $i <= 100; $i++) {
             $username = sprintf('user_%d', $i);
             $userCacheId = $this->userCacheIdGenerator->generate($username);
 
-            $this->assertFalse($this->doctrineResultCache->contains($userCacheId));
+            $this->assertFalse($this->resultCacheImplementation->contains($userCacheId));
         }
 
         $this->assertStringContainsString(
@@ -162,14 +198,14 @@ class DoctrineResultCacheWarmerCommandTest extends KernelTestCase
             $username = sprintf('user_%d', $i);
             $userCacheId = $this->userCacheIdGenerator->generate($username);
 
-            $this->assertTrue($this->doctrineResultCache->contains($userCacheId));
+            $this->assertTrue($this->resultCacheImplementation->contains($userCacheId));
         }
 
         for ($i = 61; $i <= 100; $i++) {
             $username = sprintf('user_%d', $i);
             $userCacheId = $this->userCacheIdGenerator->generate($username);
 
-            $this->assertFalse($this->doctrineResultCache->contains($userCacheId));
+            $this->assertFalse($this->resultCacheImplementation->contains($userCacheId));
         }
 
         $this->assertStringContainsString(
@@ -263,12 +299,176 @@ class DoctrineResultCacheWarmerCommandTest extends KernelTestCase
         );
     }
 
+    /**
+     * #@dataProvider provideInvalidModuloAndRemainderInputs
+     */
+    public function testItThrowsExceptionIfInvalidModuloOrRemainderParametersReceived(
+        string $expectedExceptionMessage,
+        $modulo = null,
+        $remainder = null
+    ): void {
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage($expectedExceptionMessage);
+
+        $input = [];
+
+        if (null !== $modulo) {
+            $input['--modulo'] = $modulo;
+        }
+
+        if (null !== $remainder) {
+            $input['--remainder'] = $remainder;
+        }
+
+        $this->commandTester->execute(
+            $input,
+            [
+                'capture_stderr_separately' => true,
+            ]
+        );
+    }
+
+    public function testItCanWarmUpUsersInParallelUsingModuloAndRemainderOptions(): void
+    {
+        self::assertSame(0, $this->commandTester->execute(
+            [
+                '--modulo' => 6,
+                '--remainder' => 0,
+            ],
+            [
+                'capture_stderr_separately' => true,
+            ]
+        ));
+
+        $this->assertUserCacheIsPartiallyWarmedUpForRemainders(6, 0);
+
+        self::assertSame(0, $this->commandTester->execute(
+            [
+                '--modulo' => 6,
+                '--remainder' => 1,
+            ],
+            [
+                'capture_stderr_separately' => true,
+            ]
+        ));
+
+        $this->assertUserCacheIsPartiallyWarmedUpForRemainders(6, 0, 1);
+
+        self::assertSame(0, $this->commandTester->execute(
+            [
+                '--modulo' => 6,
+                '--remainder' => 2,
+            ],
+            [
+                'capture_stderr_separately' => true,
+            ]
+        ));
+
+        $this->assertUserCacheIsPartiallyWarmedUpForRemainders(6, 0, 1, 2);
+
+        self::assertSame(0, $this->commandTester->execute(
+            [
+                '--modulo' => 6,
+                '--remainder' => 3,
+            ],
+            [
+                'capture_stderr_separately' => true,
+            ]
+        ));
+
+        $this->assertUserCacheIsPartiallyWarmedUpForRemainders(6, 0, 1, 2, 3);
+
+        self::assertSame(0, $this->commandTester->execute(
+            [
+                '--modulo' => 6,
+                '--remainder' => 4,
+            ],
+            [
+                'capture_stderr_separately' => true,
+            ]
+        ));
+
+        $this->assertUserCacheIsPartiallyWarmedUpForRemainders(6, 0, 1, 2, 3, 4);
+
+        self::assertSame(0, $this->commandTester->execute(
+            [
+                '--modulo' => 6,
+                '--remainder' => 5,
+            ],
+            [
+                'capture_stderr_separately' => true,
+            ]
+        ));
+
+        $this->assertUserCacheIsPartiallyWarmedUpForRemainders(6, 0, 1, 2, 3, 4, 5);
+    }
+
+    private function assertUserCacheIsPartiallyWarmedUpForRemainders(int $modulo, int ...$remainders): void
+    {
+        /** @var UserRepository $userRepository */
+        $userRepository = $this->getRepository(User::class);
+
+        foreach ($userRepository->findAll() as $user) {
+            $cacheId = $this->userCacheIdGenerator->generate($user->getUsername());
+
+            in_array($user->getId() % $modulo, $remainders)
+                ? self::assertTrue($this->resultCacheImplementation->contains($cacheId))
+                : self::assertFalse($this->resultCacheImplementation->contains($cacheId));
+        }
+    }
+
     public function provideInvalidFilterOption(): array
     {
         return [
             ['adsffdsfs'],
             [','],
             [',,,,,,'],
+        ];
+    }
+
+    public function provideInvalidModuloAndRemainderInputs(): array
+    {
+        return [
+            'missingRemainderOption' => [
+                'expectedExceptionMessage' => "Command option 'remainder' is expected to be specified.",
+                'modulo' => 5,
+                'remainder' => null,
+            ],
+            'missingModuloOption' => [
+                'expectedExceptionMessage' => "Command option 'modulo' is expected to be specified.",
+                'modulo' => null,
+                'remainder' => 2,
+            ],
+            'nonNumericModuloOption' => [
+                'expectedExceptionMessage' => "Command option 'modulo' is expected to be numeric.",
+                'modulo' => 'invalid',
+                'remainder' => 2,
+            ],
+            'nonNumericRemainderOption' => [
+                'expectedExceptionMessage' => "Command option 'remainder' is expected to be numeric.",
+                'modulo' => 5,
+                'remainder' => 'invalid',
+            ],
+            'tooLowModuloOption' => [
+                'expectedExceptionMessage' => "Invalid 'modulo' option received: 1, expected value: 2 <= m <= 100",
+                'modulo' => 1,
+                'remainder' => 1,
+            ],
+            'tooHighModuloOption' => [
+                'expectedExceptionMessage' => "Invalid 'modulo' option received: 101, expected value: 2 <= m <= 100",
+                'modulo' => 101,
+                'remainder' => 1,
+            ],
+            'tooLowRemainderOption' => [
+                'expectedExceptionMessage' => "Invalid 'remainder' option received: -1, expected value: 0 <= r <= 4",
+                'modulo' => 5,
+                'remainder' => -1,
+            ],
+            'tooHighRemainderOption' => [
+                'expectedExceptionMessage' => "Invalid 'remainder' option received: 5, expected value: 0 <= r <= 4",
+                'modulo' => 5,
+                'remainder' => 5,
+            ],
         ];
     }
 }